refactor: move WMTS matrix formatting into shared infrastructure (#2665)
CrossProtocolIsolationTests flagged ImageServerWmtsHandler referencing
Honua.Protocols.Ogc.Shared across protocol families. Relocate
WmtsTileMatrixFormatting next to SpatialReferenceHelpers in Honua.Hosting
(Honua.Infrastructure.Services) so both OGC Classic WMTS and GeoServices
ImageServer WMTS adapt to a neutral helper, and drop the GeoServices
project reference to Honua.Protocols.Ogc.Shared.
