+++ src/Honua.Protocols.GeoServices/Honua.Protocols.GeoServices.csproj
@@ -28,7 +28,6 @@
     <ProjectReference Include="..\Honua.Geometry\Honua.Geometry.csproj" />
     <ProjectReference Include="..\Honua.Hosting\Honua.Hosting.csproj" />
     <ProjectReference Include="..\Honua.Geoprocessing\Honua.Geoprocessing.csproj" />
-    <ProjectReference Include="..\Honua.Protocols.Ogc.Shared\Honua.Protocols.Ogc.Shared.csproj" />
     <ProjectReference Include="..\Honua.Routing\Honua.Routing.csproj" />
   </ItemGroup>
