+++ src/Honua.Protocols.GeoServices/ImageServer/Handlers/ImageServerWmtsHandler.cs
@@ -14,8 +14,8 @@ using Honua.Core.Features.Raster.Abstractions;
 using Honua.Core.Features.Raster.Domain;
 using Honua.Core.Features.Tiles;
 using Honua.Protocols.GeoServices.ImageServer.Services;
-using Honua.Protocols.Ogc.Common;
 using Honua.Infrastructure.Helpers;
+using Honua.Infrastructure.Services;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Options;
 using Microsoft.Extensions.Primitives;
